requirements.txt
aeon==1.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
boto3==1.38.8
botocore==1.38.8
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==3.5.1
Deprecated==1.2.18
dill==0.3.8
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
expecttest==0.3.0
filelock==3.13.1
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
hypothesis==6.131.9
icecream==2.1.4
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonargparse==4.39.0
kiwisolver==1.4.8
lightning==2.4.0
lightning-utilities==0.14.3
litdata==0.2.45
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.4
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytorch-lightning==1.9.5
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-pixels==3.0.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.27.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
sympy==1.13.3
termplotlib==0.3.9
texttable @ git+http://github.com/bufordtaylor/python-texttable@8cf9121154be4439ea399f0d5fa9608f9ab57403
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
tomli==2.2.1
torchmetrics==1.7.1
tqdm==4.67.1
transformers==4.51.3
triton==3.0.0
typeshed_client==2.7.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.10
xxhash==3.5.0
yarl==1.20.0
